improvement(compiler) renammed codeGen_x64 to CodeGen, now that it's archetecture independant
